+++ src/contexts/index.js
@@ -1,0 +1,17 @@
+// @flow
+
+import React from 'react';
+
+/**
+ * Copyright (c) Garuda Labs, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+// Provides the date format function to use in date fields
+// in the screen. Default to ISO string format.
+export const DateFormatContext = React.createContext<
+  (date: ?Date, format: ?string) => string,
+>(date => (date ? date.toISOString() : ''));
