+++ src/services/behaviors/index.js
@@ -9,43 +9,13 @@
  */
 
 import type {
-  BehaviorRegistry,
   Document,
   Element,
-  HvBehavior,
   Node,
   UpdateAction,
 } from 'hyperview/src/types';
 import { ACTIONS } from 'hyperview/src/types';
-import HvAlert from 'hyperview/src/behaviors/hv-alert';
-import HvCopyToClipboard from 'hyperview/src/behaviors/hv-copy-to-clipboard';
-import HvHide from 'hyperview/src/behaviors/hv-hide';
-import HvOpenSettings from 'hyperview/src/behaviors/hv-open-settings';
-import HvSetValue from 'hyperview/src/behaviors/hv-set-value';
-import HvShare from 'hyperview/src/behaviors/hv-share';
-import HvShow from 'hyperview/src/behaviors/hv-show';
-import HvToggle from 'hyperview/src/behaviors/hv-toggle';
 import { shallowCloneToRoot } from 'hyperview/src/services';
-
-const HYPERVIEW_BEHAVIORS = [
-  HvAlert,
-  HvCopyToClipboard,
-  HvHide,
-  HvOpenSettings,
-  HvSetValue,
-  HvShare,
-  HvShow,
-  HvToggle,
-];
-
-export const getRegistry = (behaviors: HvBehavior[] = []): BehaviorRegistry =>
-  [...HYPERVIEW_BEHAVIORS, ...behaviors].reduce(
-    (registry: BehaviorRegistry, behavior: HvBehavior) => ({
-      ...registry,
-      [behavior.action]: behavior,
-    }),
-    {},
-  );
 
 /**
  * The elements in the Document with the given ids will be set to
