+++ src/services/dom/helpers.js
@@ -1,0 +1,36 @@
+// @flow
+
+/**
+ * Copyright (c) Garuda Labs, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import * as Namespaces from 'hyperview/src/services/namespaces';
+import type { Document, LocalName } from 'hyperview/src/types';
+
+export const getBehaviorElements = (element: any) => {
+  // $FlowFixMe
+  const behaviorElements = Array.from(element.childNodes).filter(
+    n => n.tagName === 'behavior',
+  );
+
+  if (element.getAttribute('href') || element.getAttribute('action')) {
+    behaviorElements.unshift(element);
+  }
+
+  return behaviorElements;
+};
+
+export const getFirstTag = (document: Document, localName: LocalName) => {
+  const elements = document.getElementsByTagNameNS(
+    Namespaces.HYPERVIEW,
+    localName,
+  );
+  if (elements && elements[0]) {
+    return elements[0];
+  }
+  return null;
+};
